Celery image building
