+++ celery_worker/requirements.txt
@@ -1,24 +1,18 @@
-celery
-redis
+-e git+https://github.com/neurolearn/pynv#egg=pynv
+altair
+grabbit
 docopt
 grabbit
--e git+https://github.com/bids-standard/pybids.git@4315865#egg=pybids
--e git+https://github.com/neurolearn/pynv#egg=pynv
-duecredit
+celery
 hashids==1.2.0
-pandas
-patsy
-matplotlib
-nistats
-nilearn
-sklearn
-networkx==1.9
-nibabel
+marshmallow==2.19.1
+psycopg2==2.7.1
+redis
+pandas==0.23
+nibabel==2.1.0
 scipy
-matplotlib
-nistats
-nilearn
-fitlins
-sklearn
-tornado==5.1.1
-altair
+patsy
+pybids==0.8.0
+duecredit==0.6.1
+numpy==1.12.0
+SQLAlchemy==1.3.0
